refactor(robots): mixin for bi classes
