+++ src/lerobot/teleoperators/bi_openarm_mini/bi_openarm_mini.py
@@ -18,7 +18,8 @@ import logging
 from functools import cached_property
 
 from lerobot.types import RobotAction
-from lerobot.utils.decorators import check_if_already_connected, check_if_not_connected
+from lerobot.utils.bimanual import BimanualMixin
+from lerobot.utils.decorators import check_if_not_connected
 
 from ..openarm_mini import OpenArmMini, OpenArmMiniConfig
 from ..teleoperator import Teleoperator
@@ -27,7 +28,7 @@ from .config_bi_openarm_mini import BiOpenArmMiniConfig
 logger = logging.getLogger(__name__)
 
 
-class BiOpenArmMini(Teleoperator):
+class BiOpenArmMini(BimanualMixin, Teleoperator):
     """Bimanual OpenArm Mini teleoperator.
 
     Composes two single-arm :class:`OpenArmMini` instances. Action and feedback
@@ -77,27 +78,6 @@ class BiOpenArmMini(Teleoperator):
             **{f"right_{k}": v for k, v in self.right_arm.feedback_features.items()},
         }
 
-    @property
-    def is_connected(self) -> bool:
-        return self.left_arm.is_connected and self.right_arm.is_connected
-
-    @check_if_already_connected
-    def connect(self, calibrate: bool = True) -> None:
-        self.left_arm.connect(calibrate)
-        self.right_arm.connect(calibrate)
-
-    @property
-    def is_calibrated(self) -> bool:
-        return self.left_arm.is_calibrated and self.right_arm.is_calibrated
-
-    def calibrate(self) -> None:
-        self.left_arm.calibrate()
-        self.right_arm.calibrate()
-
-    def configure(self) -> None:
-        self.left_arm.configure()
-        self.right_arm.configure()
-
     def setup_motors(self) -> None:
         self.left_arm.setup_motors()
         self.right_arm.setup_motors()
@@ -119,8 +99,3 @@ class BiOpenArmMini(Teleoperator):
             self.left_arm.send_feedback(left_fb)
         if right_fb:
             self.right_arm.send_feedback(right_fb)
-
-    @check_if_not_connected
-    def disconnect(self) -> None:
-        self.left_arm.disconnect()
-        self.right_arm.disconnect()
